# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COW Sites Hajj Monitoring API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SiteStatus = (typeof SiteStatus)[keyof typeof SiteStatus];

export const SiteStatus = {
  operational: "operational",
  degraded: "degraded",
  offline: "offline",
} as const;

export interface Site {
  id: number;
  name: string;
  location: string;
  zone: string;
  status: SiteStatus;
  latitude?: number;
  longitude?: number;
  deployedAt?: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateSiteBodyStatus =
  (typeof CreateSiteBodyStatus)[keyof typeof CreateSiteBodyStatus];

export const CreateSiteBodyStatus = {
  operational: "operational",
  degraded: "degraded",
  offline: "offline",
} as const;

export interface CreateSiteBody {
  name: string;
  location: string;
  zone: string;
  status?: CreateSiteBodyStatus;
  latitude?: number;
  longitude?: number;
  deployedAt?: string;
}

export type UpdateSiteBodyStatus =
  (typeof UpdateSiteBodyStatus)[keyof typeof UpdateSiteBodyStatus];

export const UpdateSiteBodyStatus = {
  operational: "operational",
  degraded: "degraded",
  offline: "offline",
} as const;

export interface UpdateSiteBody {
  name?: string;
  location?: string;
  zone?: string;
  status?: UpdateSiteBodyStatus;
  latitude?: number;
  longitude?: number;
}

export type TicketType = (typeof TicketType)[keyof typeof TicketType];

export const TicketType = {
  power: "power",
  telecom: "telecom",
} as const;

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  open: "open",
  in_progress: "in_progress",
  resolved: "resolved",
  closed: "closed",
} as const;

export type TicketPriority =
  (typeof TicketPriority)[keyof typeof TicketPriority];

export const TicketPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Ticket {
  id: number;
  siteId: number;
  siteName: string;
  type: TicketType;
  status: TicketStatus;
  priority: TicketPriority;
  title: string;
  description?: string;
  assignedTo?: string;
  resolvedAt?: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateTicketBodyType =
  (typeof CreateTicketBodyType)[keyof typeof CreateTicketBodyType];

export const CreateTicketBodyType = {
  power: "power",
  telecom: "telecom",
} as const;

export type CreateTicketBodyPriority =
  (typeof CreateTicketBodyPriority)[keyof typeof CreateTicketBodyPriority];

export const CreateTicketBodyPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface CreateTicketBody {
  siteId: number;
  type: CreateTicketBodyType;
  priority: CreateTicketBodyPriority;
  title: string;
  description?: string;
  assignedTo?: string;
}

export type UpdateTicketBodyStatus =
  (typeof UpdateTicketBodyStatus)[keyof typeof UpdateTicketBodyStatus];

export const UpdateTicketBodyStatus = {
  open: "open",
  in_progress: "in_progress",
  resolved: "resolved",
  closed: "closed",
} as const;

export type UpdateTicketBodyPriority =
  (typeof UpdateTicketBodyPriority)[keyof typeof UpdateTicketBodyPriority];

export const UpdateTicketBodyPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface UpdateTicketBody {
  status?: UpdateTicketBodyStatus;
  priority?: UpdateTicketBodyPriority;
  assignedTo?: string;
  description?: string;
  resolvedAt?: string;
}

export interface DashboardSummary {
  totalSites: number;
  operationalSites: number;
  degradedSites: number;
  offlineSites: number;
  totalTickets: number;
  openTickets: number;
  inProgressTickets: number;
  resolvedTickets: number;
  criticalTickets: number;
  powerOutageTickets: number;
  telecomOutageTickets: number;
  avgResolutionHours: number;
}

export interface OutagesByDay {
  date: string;
  powerOutages: number;
  telecomOutages: number;
}

export interface OutagesByType {
  type: string;
  count: number;
  percentage: number;
}

export type SiteStatusSummaryStatus =
  (typeof SiteStatusSummaryStatus)[keyof typeof SiteStatusSummaryStatus];

export const SiteStatusSummaryStatus = {
  operational: "operational",
  degraded: "degraded",
  offline: "offline",
} as const;

export interface SiteStatusSummary {
  siteId: number;
  siteName: string;
  zone: string;
  status: SiteStatusSummaryStatus;
  openTickets: number;
  lastIncident?: string;
}

export type RecentActivityType =
  (typeof RecentActivityType)[keyof typeof RecentActivityType];

export const RecentActivityType = {
  power: "power",
  telecom: "telecom",
} as const;

export type RecentActivityStatus =
  (typeof RecentActivityStatus)[keyof typeof RecentActivityStatus];

export const RecentActivityStatus = {
  open: "open",
  in_progress: "in_progress",
  resolved: "resolved",
  closed: "closed",
} as const;

export type RecentActivityPriority =
  (typeof RecentActivityPriority)[keyof typeof RecentActivityPriority];

export const RecentActivityPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface RecentActivity {
  ticketId: number;
  ticketTitle: string;
  siteName: string;
  type: RecentActivityType;
  status: RecentActivityStatus;
  priority: RecentActivityPriority;
  timestamp: string;
  action: string;
}

export type ListTicketsParams = {
  siteId?: number;
  type?: ListTicketsType;
  status?: ListTicketsStatus;
  priority?: ListTicketsPriority;
};

export type ListTicketsType =
  (typeof ListTicketsType)[keyof typeof ListTicketsType];

export const ListTicketsType = {
  power: "power",
  telecom: "telecom",
} as const;

export type ListTicketsStatus =
  (typeof ListTicketsStatus)[keyof typeof ListTicketsStatus];

export const ListTicketsStatus = {
  open: "open",
  in_progress: "in_progress",
  resolved: "resolved",
  closed: "closed",
} as const;

export type ListTicketsPriority =
  (typeof ListTicketsPriority)[keyof typeof ListTicketsPriority];

export const ListTicketsPriority = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;