# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * COW Sites Hajj Monitoring API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all COW sites
 */
export const ListSitesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  zone: zod.string(),
  status: zod.enum(["operational", "degraded", "offline"]),
  latitude: zod.number().optional(),
  longitude: zod.number().optional(),
  deployedAt: zod.coerce.date().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListSitesResponse = zod.array(ListSitesResponseItem);

/**
 * @summary Create a new COW site
 */
export const CreateSiteBody = zod.object({
  name: zod.string(),
  location: zod.string(),
  zone: zod.string(),
  status: zod.enum(["operational", "degraded", "offline"]).optional(),
  latitude: zod.number().optional(),
  longitude: zod.number().optional(),
  deployedAt: zod.coerce.date().optional(),
});

/**
 * @summary Get a COW site by ID
 */
export const GetSiteParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSiteResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  zone: zod.string(),
  status: zod.enum(["operational", "degraded", "offline"]),
  latitude: zod.number().optional(),
  longitude: zod.number().optional(),
  deployedAt: zod.coerce.date().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a COW site
 */
export const UpdateSiteParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSiteBody = zod.object({
  name: zod.string().optional(),
  location: zod.string().optional(),
  zone: zod.string().optional(),
  status: zod.enum(["operational", "degraded", "offline"]).optional(),
  latitude: zod.number().optional(),
  longitude: zod.number().optional(),
});

export const UpdateSiteResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  zone: zod.string(),
  status: zod.enum(["operational", "degraded", "offline"]),
  latitude: zod.number().optional(),
  longitude: zod.number().optional(),
  deployedAt: zod.coerce.date().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary List all outage tickets
 */
export const ListTicketsQueryParams = zod.object({
  siteId: zod.coerce.number().optional(),
  type: zod.enum(["power", "telecom"]).optional(),
  status: zod.enum(["open", "in_progress", "resolved", "closed"]).optional(),
  priority: zod.enum(["critical", "high", "medium", "low"]).optional(),
});

export const ListTicketsResponseItem = zod.object({
  id: zod.number(),
  siteId: zod.number(),
  siteName: zod.string(),
  type: zod.enum(["power", "telecom"]),
  status: zod.enum(["open", "in_progress", "resolved", "closed"]),
  priority: zod.enum(["critical", "high", "medium", "low"]),
  title: zod.string(),
  description: zod.string().optional(),
  assignedTo: zod.string().optional(),
  resolvedAt: zod.coerce.date().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListTicketsResponse = zod.array(ListTicketsResponseItem);

/**
 * @summary Create a new outage ticket
 */
export const CreateTicketBody = zod.object({
  siteId: zod.number(),
  type: zod.enum(["power", "telecom"]),
  priority: zod.enum(["critical", "high", "medium", "low"]),
  title: zod.string(),
  description: zod.string().optional(),
  assignedTo: zod.string().optional(),
});

/**
 * @summary Get a ticket by ID
 */
export const GetTicketParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTicketResponse = zod.object({
  id: zod.number(),
  siteId: zod.number(),
  siteName: zod.string(),
  type: zod.enum(["power", "telecom"]),
  status: zod.enum(["open", "in_progress", "resolved", "closed"]),
  priority: zod.enum(["critical", "high", "medium", "low"]),
  title: zod.string(),
  description: zod.string().optional(),
  assignedTo: zod.string().optional(),
  resolvedAt: zod.coerce.date().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a ticket
 */
export const UpdateTicketParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTicketBody = zod.object({
  status: zod.enum(["open", "in_progress", "resolved", "closed"]).optional(),
  priority: zod.enum(["critical", "high", "medium", "low"]).optional(),
  assignedTo: zod.string().optional(),
  description: zod.string().optional(),
  resolvedAt: zod.coerce.date().optional(),
});

export const UpdateTicketResponse = zod.object({
  id: zod.number(),
  siteId: zod.number(),
  siteName: zod.string(),
  type: zod.enum(["power", "telecom"]),
  status: zod.enum(["open", "in_progress", "resolved", "closed"]),
  priority: zod.enum(["critical", "high", "medium", "low"]),
  title: zod.string(),
  description: zod.string().optional(),
  assignedTo: zod.string().optional(),
  resolvedAt: zod.coerce.date().optional(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Get dashboard KPI summary
 */
export const GetDashboardSummaryResponse = zod.object({
  totalSites: zod.number(),
  operationalSites: zod.number(),
  degradedSites: zod.number(),
  offlineSites: zod.number(),
  totalTickets: zod.number(),
  openTickets: zod.number(),
  inProgressTickets: zod.number(),
  resolvedTickets: zod.number(),
  criticalTickets: zod.number(),
  powerOutageTickets: zod.number(),
  telecomOutageTickets: zod.number(),
  avgResolutionHours: zod.number(),
});

/**
 * @summary Get outage ticket counts grouped by day
 */
export const GetOutagesByDayResponseItem = zod.object({
  date: zod.string(),
  powerOutages: zod.number(),
  telecomOutages: zod.number(),
});
export const GetOutagesByDayResponse = zod.array(GetOutagesByDayResponseItem);

/**
 * @summary Get outage counts by type (power vs telecom)
 */
export const GetOutagesByTypeResponseItem = zod.object({
  type: zod.string(),
  count: zod.number(),
  percentage: zod.number(),
});
export const GetOutagesByTypeResponse = zod.array(GetOutagesByTypeResponseItem);

/**
 * @summary Get status breakdown of all COW sites
 */
export const GetSiteStatusResponseItem = zod.object({
  siteId: zod.number(),
  siteName: zod.string(),
  zone: zod.string(),
  status: zod.enum(["operational", "degraded", "offline"]),
  openTickets: zod.number(),
  lastIncident: zod.coerce.date().optional(),
});
export const GetSiteStatusResponse = zod.array(GetSiteStatusResponseItem);

/**
 * @summary Get recent ticket activity feed
 */
export const GetRecentActivityResponseItem = zod.object({
  ticketId: zod.number(),
  ticketTitle: zod.string(),
  siteName: zod.string(),
  type: zod.enum(["power", "telecom"]),
  status: zod.enum(["open", "in_progress", "resolved", "closed"]),
  priority: zod.enum(["critical", "high", "medium", "low"]),
  timestamp: zod.coerce.date(),
  action: zod.string(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);